feat: optional auto-open disclosure triangle for PR comment detail views
This option will allow teams who prefer it, to draw more attention
to the short resource diff and/or the detailed log views in the
PR comment, to highlight potentially destructive changes more readily.

The default behavior of these new optional inputs matches the prior
behavior, so there is no change for teams who do not opt in.