fix: change counting logic for case of zero changes
Prior to this PR,

In the case where a `tf.diff.txt` is generated, but it contains
no changes (e.g. only moves and imports), then the `grep` that
looks for lines not starting with `^# ` would find no lines at all,
causing `grep` to return a non-zero exit code, which causes
the CI/CD workflow to fail.

This PR fixes the issue by adding `|| true` to the relevant `grep`.
The `true` command will output zero lines with a zero exit code,
so it makes a good fallback here when the `grep` fails
(because the `wc --lines` that follows will count zero lines).